refactor: move UserProfileRoute to routes directory and update imports
